server/requirements.txt
flask==1.1.2
flask-sqlalchemy==2.5.1
sqlalchemy==1.3.24
flask-login==0.5.0
flask-cors==3.0.10
beautifulsoup4==4.9.3
flask_wtf==0.14.3
bleach==3.3.0
greenlet==0.4.17
gunicorn
Pillow==8.2.0
pyyaml
Flask-Limiter